requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
django==5.2.8
django-constance==4.3.4
django-ninja==1.5.0
django-unfold==0.72.0
environs==14.5.0
idna==3.11
marshmallow==4.1.0
pillow==12.0.0
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
sqlparse==0.5.4
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
